requirements.txt
Django==5.2.4
django-admin-sortable2==2.2.8
django-tinymce==4.1.0
requests==2.32.4
environs==14.2.0